# Conflicts:
#	crates/bevy_wgpu/src/lib.rs
#	crates/bevy_wgpu/src/wgpu_renderer.rs

+++ crates/bevy_input/src/mouse.rs
@@ -1,5 +1,4 @@
-use super::keyboard::ElementState;
-use crate::Input;
+use crate::{ElementState, Input};
 use bevy_app::prelude::{EventReader, Events};
 use bevy_ecs::{Local, Res, ResMut};
 use bevy_math::Vec2;
@@ -18,7 +17,7 @@ pub enum MouseButton {
     Left,
     Right,
     Middle,
-    Other(u8),
+    Other(u16),
 }
 
 /// A mouse motion event
